% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obj.R
\name{get_obj}
\alias{get_obj}
\title{plot PCA}
\usage{
get_obj(x, y)
}
\arguments{
\item{x}{experimental design input}

\item{y}{protein groups input}
}
\description{
This function loads files and plot PCA as output
}
