% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntensityMarix_process.R
\name{IntensityMarix_process}
\alias{IntensityMarix_process}
\title{Process Marix Intensity}
\usage{
IntensityMarix_process(IntensityMarix, threshold = 1, Imputation = TRUE,
  Normalize_columns = TRUE, Normalize_rows = TRUE)
}
\arguments{
\item{IntensityMarix}{data_matrix as input}

\item{threshold}{the default value is set to 1}

\item{Imputation}{the default value is set to TRUE}

\item{Normalize_columns}{the default value is set to TRUE}

\item{Normalize_rows}{the default value is set to TRUE}
}
\description{
Process Marix Intensity
}
