% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_ggboxplot.R
\name{matrix_ggboxplot}
\alias{matrix_ggboxplot}
\title{ggbox plot}
\usage{
matrix_ggboxplot(data_matrix, xlabel = "Samples", ylabel = "Value",
  maintitle = "Distribution")
}
\arguments{
\item{data_matrix}{data_matrix as input}

\item{xlabel}{label for x}

\item{ylabel}{label for y}

\item{maintitle}{title for main}
}
\description{
ggbox plot
}
