% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingvalue_filtering.R
\name{missingvalue_filtering}
\alias{missingvalue_filtering}
\title{missing value filtering function}
\usage{
missingvalue_filtering(data.matrix, threshold = 1)
}
\arguments{
\item{threshold}{the value for threshold is set to 1as default}

\item{data_matrix}{data_matrix as input}
}
\description{
missing value filtering function
}
