% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot.R
\name{PCA_plot}
\alias{PCA_plot}
\title{PCA screen plot original}
\usage{
PCA_plot(data_matrix, grouping)
}
\arguments{
\item{data_matrix}{take data_matrix as input}

\item{grouping}{optional for plotting}
}
\description{
PCA screen plot original
}
