% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Screenplot.R
\name{PCA_Screeplot}
\alias{PCA_Screeplot}
\title{PCA screen plot version one}
\usage{
PCA_Screeplot(data_matrix)
}
\arguments{
\item{data_matrix}{take data_matrix as input}
}
\description{
PCA screen plot version one
}
