% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Screenplot_2.R
\name{PCA_Screeplot_2}
\alias{PCA_Screeplot_2}
\title{PCA screen plot version two}
\usage{
PCA_Screeplot_2(prcomp.out)
}
\arguments{
\item{prcomp.out}{generate output}
}
\description{
PCA screen plot version two
}
