% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_test_plot.R
\name{pca_test_plot}
\alias{pca_test_plot}
\title{plot PCA}
\usage{
pca_test_plot(expinput, proteininput)
}
\arguments{
\item{expinput}{experimental design input}

\item{proteininput}{protein groups input}
}
\description{
This function loads files and plot PCA as output
}
