% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one.R
\name{plot_one}
\alias{plot_one}
\title{plot PCA}
\usage{
plot_one(expinput, proteininput)
}
\arguments{
\item{expinput}{experimental design input}

\item{proteininput}{protein groups input}
}
\description{
This function loads files and plot PCA as output
}
