% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one.R
\name{plot_one}
\alias{plot_one}
\title{plot PCA}
\usage{
plot_one(newinput1, newinput2)
}
\arguments{
\item{newinput1}{experimental design input}

\item{newinput2}{protein groups input}
}
\description{
This function loads files and plot PCA as output
}
