% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectdata_From_ProteinGroupsFile.R
\name{Selectdata_From_ProteinGroupsFile}
\alias{Selectdata_From_ProteinGroupsFile}
\title{Select data from protein groups}
\usage{
Selectdata_From_ProteinGroupsFile(proteinfileinput,
  expression_data_tobe_selected)
}
\arguments{
\item{proteinfileinput}{obtained from file input}

\item{expression_data_tobe_selected}{set a default selection scale}
}
\description{
Select data from protein groups
}
