% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectdata_From_ProteinGroupsFile.R
\name{Selectdata_From_ProteinGroupsFile}
\alias{Selectdata_From_ProteinGroupsFile}
\title{Select data from protein groups}
\usage{
Selectdata_From_ProteinGroupsFile(proteinGroups,
  expression_data_tobe_selected = "LFQ_intensity")
}
\arguments{
\item{proteinGroups}{obtained from file input}
}
\description{
Select data from protein groups
}
